Reduce calls to malloc in unordered_multimap (#97)
Reduce the number of malloc calls in unordered_multimap. Improves
efficiency by 60% (meaning it takes less than half the time that it used to).

+++ containers.h
@@ -276,7 +276,7 @@ unordered_multimap_init(size_t key_size,
 
 /* Utility */
 int unordered_multimap_rehash(unordered_multimap me);
-int unordered_multimap_size(unordered_multimap me);
+size_t unordered_multimap_size(unordered_multimap me);
 int unordered_multimap_is_empty(unordered_multimap me);
 
 /* Accessing */
